+++ R/00-NAMESPACE.R
@@ -1,12 +1,12 @@
 #' @import shiny
 #' @importFrom car durbinWatsonTest vif
 #' @importFrom coda raftery.diag gelman.diag geweke.diag heidel.diag mcmc
-#' @importFrom BMSC constrSelEst createFormula
+#' @importFrom DataTools downloadModelUI downloadModelServer importDataUI importDataServer
+#'  remoteModelsUI remoteModelsServer tryCatchWithWarningsAndErrors uploadModelUI uploadModelServer
 #' @importFrom DT datatable renderDT DTOutput
 #' @importFrom dplyr ungroup summarise group_by "%>%" summarise_ group_by_
 #' @importFrom ggplot2 aes_string element_text geom_errorbar geom_point ggplot
 #' labs scale_x_discrete theme aes aes_ geom_boxplot ylab xlab
-#' @importFrom DataTools importDataUI importDataServer tryCatchWithWarningsAndErrors
 #' @importFrom shinyjs useShinyjs alert
 #' @importFrom shinyWidgets pickerInput updatePickerInput
 #' @importFrom openxlsx write.xlsx
@@ -17,9 +17,10 @@
 #' @importFrom stats lm as.formula predict rexp rnorm terms rpois median quantile sd qnorm chisq.test cor lm.fit model.matrix na.omit
 #' @importFrom jsonlite toJSON
 #' @importFrom grDevices dev.off pdf png svg tiff
-#' @importFrom utils capture.output head read.csv write.table
+#' @importFrom utils capture.output head packageVersion read.csv write.table
 #' @importFrom tidyr gather
-#' @importFrom BMSC prepModelNames get_avg_model getModelFits get_model_weights plotModelFit
+#' @importFrom BMSC constrSelEst createFormula prepModelNames get_avg_model getModelFits 
+#'  get_model_weights plotModelFit
 NULL
 
 #' Server and UI Functions for Shiny Module